+++ components/atoms/ALink.js
@@ -2,7 +2,7 @@ import qs from 'qs'
 import get from 'lodash/get'
 export default {
   functional: true,
-  name: 'AppLink',
+  name: 'ALink',
   props: {
     to: {
       type: [String, Object],
@@ -45,7 +45,7 @@ export default {
        * Fix absolute URLs which really should have been relative
        * Example: "http://stamford.localhost/home" -> "/home"
        */
-      const hostUrl = process.env.APPLINK_RELATIVE_BASE_URL
+      const hostUrl = process.env.RELATIVE_BASE_URL
       const indexOfHost = url.indexOf('//' + hostUrl)
       if (indexOfHost !== -1) {
         url = url.substring(indexOfHost + 2 + hostUrl.length)