+++ components/pages/EditorialCategory.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="EditorialCategory">
+    EditorialCategory
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'EditorialCategory',
+  }
+</script>