+++ components/pages/GalleryArchive.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="GalleryArchive">
+    GalleryArchive
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'GalleryArchive',
+  }
+</script>