+++ components/pages/GallerySingle.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="GallerySingle">
+    GallerySingle
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'GallerySingle',
+  }
+</script>