+++ components/pages/Home/HomeTopStories.vue
@@ -1,0 +1,24 @@
+<template>
+  <div
+    class="HomeTopStories // lg:grid lg:grid-rows-3 lg:grid-flow-col lg:gap-4"
+  >
+    <div class="p-4 my-4 bg-brand-teal-200 lg:row-span-3 lg:col-span-2 lg:my-0">
+      In vel ultrices nisl
+    </div>
+    <div class="p-4 my-4 bg-brand-teal-200 lg:row-span-2 lg:col-span-1 lg:my-0">
+      In vel ultrices nisl
+    </div>
+    <div class="p-4 my-4 bg-brand-teal-200 lg:row-span-1 lg:col-span-1 lg:my-0">
+      In vel ultrices nisl
+    </div>
+  </div>
+</template>
+
+<script>
+  // import gql from 'graphql-tag'
+
+  export default {
+    name: 'HomeTopStories',
+    // apollo: {},
+  }
+</script>