+++ components/pages/ListArchive.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="ListArchive">
+    ListArchive
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'ListArchive',
+  }
+</script>