+++ components/pages/ListSingle.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="ListSingle">
+    ListSingle
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'ListSingle',
+  }
+</script>