+++ components/pages/Page.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="Page">
+    Page
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'Page',
+  }
+</script>