+++ components/pages/VerticalCategory.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="VerticalCategory">
+    VerticalCategory
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'VerticalCategory',
+  }
+</script>