+++ components/pages/VideoArchive.vue
@@ -1,0 +1,11 @@
+<template>
+  <div class="VideoArchive">
+    VideoArchive
+  </div>
+</template>
+
+<script>
+  export default {
+    name: 'VideoArchive',
+  }
+</script>