+++ pages/gallery.js
@@ -1,0 +1,2 @@
+import GalleryArchive from '~/components/pages/GalleryArchive'
+export default GalleryArchive