+++ pages/list.js
@@ -1,0 +1,2 @@
+import ListArchive from '~/components/pages/ListArchive'
+export default ListArchive