+++ pages/videos.js
@@ -1,0 +1,2 @@
+import VideoArchive from '~/components/pages/VideoArchive'
+export default VideoArchive